Make CLI server and ui commands to work with db driver URIs (#1411)
#1297 & #1374 added support for using non-default drivers for database URIs, but they do not work with the CLI commands mlflow server and mlflow ui. Adding support here.
